+++ R/broom.helpers-package.R
@@ -49,7 +49,7 @@ utils::globalVariables(c("."))
     dplyr::select(
       dplyr::any_of(
         c(
-          "group_by", "y.level", "component", "term", "original_term",
+          "group_by", "y.level", "component", "tau", "term", "original_term",
           "variable", "instrumental", "var_label", "var_class", "var_type",
           "var_nlevels", "header_row", "contrasts", "contrasts_type",
           "reference_row", "label", "n_obs", "n_ind", "n_event", "exposure"