+++ NEWS.md
@@ -1,5 +1,10 @@
 # broom.helpers (development version)
 
+**New supported models**
+
+- support for `quantreg::rq()` models, use `tidy_with_broom_or_parameters()`,
+  default tidier of `tidy_plus_plus()` (#299)
+
 # broom.helpers 1.21.0
 
 **New supported models**